+++ src/core/utils.jl
@@ -550,7 +550,7 @@ function get_monthly_time_steps(year::Int; time_steps_per_hour=1)
     for m in range(1, stop=12)
         n_days = daysinmonth(Date(string(year) * "-" * string(m)))
         stop = n_days * 24 * time_steps_per_hour + i - 1
-        if m == 2 && isleapyear(year)
+        if m == 12 && isleapyear(year)
             stop -= 24 * time_steps_per_hour  # TODO support extra day in leap years?
         end
         steps = [step for step in range(i, stop=stop)]